feat replace _id to id

+++ src/modules/genre/genre.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { Genre } from './models/genre';
 import { CreateGenreInput } from './input/create-genre.input';
 import { UpdateGenreInput } from './input/update-genreinput';
@@ -13,9 +13,8 @@ export class GenreService {
   constructor(private readonly httpService: HttpService) {}
   private genres: Genre[] = [];
 
-  public async createGenre(
-    createGenre: CreateGenreInput,
-  ): Promise<Genre> {
+  public async createGenre(createGenre: CreateGenreInput): Promise<Genre> {
+    if (!process.env.token) throw new ForbiddenException();
     const data = await this.httpService.axiosRef.post(
       'http://localhost:3001/v1/genres',
       createGenre,
@@ -26,14 +25,14 @@ export class GenreService {
       },
     );
 
-    return data.data;
+    return this.replaceId(data.data);
   }
 
-  public async updateGenre(
-    updateGenre: UpdateGenreInput,
-  ): Promise<Genre> {
+  public async updateGenre(updateGenre: UpdateGenreInput): Promise<Genre> {
+    if (!process.env.token) throw new ForbiddenException();
+
     const data = await this.httpService.axiosRef.put(
-      'http://localhost:3001/v1/genres/' + updateGenre._id,
+      'http://localhost:3001/v1/genres/' + updateGenre.id,
       updateGenre,
       {
         headers: {
@@ -42,16 +41,16 @@ export class GenreService {
       },
     );
 
-    return data.data;
+    return this.replaceId(data.data);
   }
   public async getGenre(
     getGenretArg: GetGenreArg,
-  ): Promise<AxiosResponse<Genre>> {            
+  ): Promise<AxiosResponse<Genre>> {
     const data = await this.httpService.axiosRef.get(
-      'http://localhost:3001/v1/genres/' + getGenretArg._id,
+      'http://localhost:3001/v1/genres/' + getGenretArg.id,
     );
-      
-    return data.data;
+
+    return this.replaceId(data.data);
   }
   public async getGenres(
     getGenretArg: GetGenresArg,
@@ -60,11 +59,18 @@ export class GenreService {
       'http://localhost:3001/v1/genres',
     );
 
-    return data.data.items;
+    return data.data.items.map((item) => {
+      this.replaceId(item);
+
+      return item;
+    });
   }
-  public async deleteGenre(getGenretArg: DeleteGenreInput): Promise<DeleteGenreInput> {
+  public async deleteGenre(
+    getGenretArg: DeleteGenreInput,
+  ): Promise<DeleteGenreInput> {
+    if (!process.env.token) throw new ForbiddenException();
     const data = await this.httpService.axiosRef.delete(
-      'http://localhost:3001/v1/genres/' + getGenretArg._id,
+      'http://localhost:3001/v1/genres/' + getGenretArg.id,
       {
         headers: {
           Authorization: `Token ${process.env.token}`,
@@ -73,5 +79,12 @@ export class GenreService {
     );
 
     return getGenretArg;
+  }
+
+  replaceId(data) {
+    if (!data._id) return data;
+    data.id = data._id;
+    delete data._id;
+    return data;
   }
 }
